feat(DS-S2): 로그아웃 ProfileMenu 드롭다운으로 통합 + 중복 제거
- profile-menu.tsx: 로그아웃 항목 활성화 (logoutUser + router.push)
- app-sidebar.tsx: SidebarFooter LogoutButton 제거
- dashboard/page.tsx: 헤더 TopBarSlot LogoutButton 제거
- logout-button.tsx: 파일 삭제 (모든 사용처 정리 완료)
